+++ action/test/run.test.ts
@@ -1,7 +1,7 @@
 import {
   VISUAL_REGRESSION_CONTEXT,
   VISUAL_TESTS_FAILED_TO_EXECUTE
-} from 'shared';
+} from 'shared/constants';
 import { afterEach, beforeEach, describe, expect, it, mock } from 'bun:test';
 import { EventEmitter } from 'events';
 import { Readable } from 'stream';
@@ -77,10 +77,29 @@ const listAllObjects = async (
     ...(await listAllObjects(input, response.NextContinuationToken))
   ];
 };
-mock.module('shared/s3Client', () => ({
+async function getKeysFromS3(directory: string, hash: string, bucket: string) {
+  const allContents = await listAllObjects({
+    Bucket: bucket,
+    Prefix: `${directory}/${hash}/`
+  });
+  const keys = allContents.map(
+    (content: { Key?: string }) => content.Key ?? ''
+  );
+  return keys.filter(
+    (path: string) => path && !path.includes('actions-runner')
+  );
+}
+mock.module('shared/s3', () => ({
   s3Client: {},
   listObjects: listObjectsMock,
   listAllObjects,
+  getKeysFromS3,
+  filterNewImages: mock(),
+  toBaseImagePath: mock(),
+  getBaseImagePaths: mock(),
+  getBaseImagePathsFromOriginal: mock(),
+  encodeS3CopySource: mock(),
+  updateBaseImages: mock(),
   getObject: getObjectMock,
   putObject: putObjectMock,
   copyObject: mock()
