+++ app/backend/src/getTemporaryObjectUrl.ts
@@ -1,4 +1,4 @@
-import { s3Client } from 'shared';
+import { s3Client } from 'shared/s3';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import { GetObjectCommand } from '@aws-sdk/client-s3';
