+++ app/backend/test/fetchCurrentPage.test.ts
@@ -1,6 +1,6 @@
 import { fetchCurrentPage } from '../src/fetchCurrentPage';
 import { describe, expect, it, mock } from 'bun:test';
-import { NEW_IMAGES_DIRECTORY } from 'shared';
+import { NEW_IMAGES_DIRECTORY } from 'shared/constants';
 
 mock.module('../src/getTemporaryObjectUrl', () => ({
   getTemporaryObjectUrl: mock(() => 'url')
@@ -16,14 +16,34 @@ const listObjectsMock = mock(() => ({
     { Key: `${pathPrefix}/LARGE/invalidPage/new.png` }
   ]
 }));
-mock.module('shared/s3Client', () => ({
+mock.module('shared/s3', () => ({
   s3Client: {},
   listObjects: listObjectsMock,
   listAllObjects,
+  getKeysFromS3,
+  filterNewImages: mock(),
+  toBaseImagePath: mock(),
+  getBaseImagePaths: mock(),
+  getBaseImagePathsFromOriginal: mock(),
+  encodeS3CopySource: mock(),
+  updateBaseImages: mock(),
   getObject: mock(),
   putObject: mock(),
   copyObject: mock()
 }));
+
+async function getKeysFromS3(directory: string, hash: string, bucket: string) {
+  const allContents = await listAllObjects({
+    Bucket: bucket,
+    Prefix: `${directory}/${hash}/`
+  });
+  const keys = allContents.map(
+    (content: { Key?: string }) => content.Key ?? ''
+  );
+  return keys.filter(
+    (path: string) => path && !path.includes('actions-runner')
+  );
+}
 
 async function listAllObjects(
   input: { Bucket: string; Prefix: string; ContinuationToken?: string },
