+++ shared/constants.ts
@@ -1,0 +1,9 @@
+export const VISUAL_REGRESSION_CONTEXT = 'Visual Regression';
+export const DIFF_IMAGE_NAME = 'diff';
+export const VISUAL_TESTS_FAILED_TO_EXECUTE =
+  'Visual tests failed to execute successfully.';
+export const BASE_IMAGES_DIRECTORY = 'base-images';
+export const BASE_IMAGE_NAME = 'base';
+export const NEW_IMAGE_NAME = 'new';
+export const NEW_IMAGES_DIRECTORY = 'new-images';
+export const ORIGINAL_NEW_IMAGES_DIRECTORY = 'original-new-images';
