--- shared/index.ts
@@ -1,8 +1,0 @@
-export * from './s3Client';
-
-export const VISUAL_REGRESSION_CONTEXT = 'Visual Regression';
-export const NEW_IMAGES_DIRECTORY = 'new-images';
-export const ORIGINAL_NEW_IMAGES_DIRECTORY = 'original-new-images';
-export const DIFF_IMAGE_NAME = 'diff';
-export const VISUAL_TESTS_FAILED_TO_EXECUTE =
-  'Visual tests failed to execute successfully.';
