--- shared/s3Client.ts
@@ -1,75 +1,0 @@
-import {
-  S3Client,
-  ListObjectsV2Command,
-  ListObjectsV2CommandInput,
-  ListObjectsV2CommandOutput,
-  GetObjectCommand,
-  GetObjectCommandInput,
-  GetObjectCommandOutput,
-  PutObjectCommand,
-  PutObjectCommandInput,
-  PutObjectCommandOutput,
-  CopyObjectCommand,
-  CopyObjectCommandInput,
-  CopyObjectCommandOutput
-} from '@aws-sdk/client-s3';
-
-export const s3Client = new S3Client();
-
-export const listObjects = (
-  input: ListObjectsV2CommandInput
-): Promise<ListObjectsV2CommandOutput> =>
-  s3Client.send(new ListObjectsV2Command(input));
-
-export const listAllObjects = async (
-  input: Omit<ListObjectsV2CommandInput, 'ContinuationToken'>,
-  continuationToken?: string
-): Promise<NonNullable<ListObjectsV2CommandOutput['Contents']>> => {
-  const response = await listObjects({
-    ...input,
-    ...(continuationToken && { ContinuationToken: continuationToken })
-  });
-  const contents = response.Contents ?? [];
-  if (!response.IsTruncated) return contents;
-  const remaining = await listAllObjects(input, response.NextContinuationToken);
-  return [...contents, ...remaining];
-};
-
-export const getObject = (
-  input: GetObjectCommandInput
-): Promise<GetObjectCommandOutput> =>
-  s3Client.send(new GetObjectCommand(input));
-
-export const putObject = (
-  input: PutObjectCommandInput
-): Promise<PutObjectCommandOutput> =>
-  s3Client.send(new PutObjectCommand(input));
-
-export const copyObject = (
-  input: CopyObjectCommandInput
-): Promise<CopyObjectCommandOutput> =>
-  s3Client.send(new CopyObjectCommand(input));
-
-export const BASE_IMAGES_DIRECTORY = 'base-images';
-export const BASE_IMAGE_NAME = 'base';
-export const NEW_IMAGE_NAME = 'new';
-
-export function encodeS3CopySource(bucket: string, key: string): string {
-  return `${bucket}/${key.split('/').map(encodeURIComponent).join('/')}`;
-}
-
-export function filterNewImages(s3Paths: string[]): string[] {
-  return s3Paths.filter(path =>
-    path.match(new RegExp(`/${NEW_IMAGE_NAME}.png`))
-  );
-}
-
-export function toBaseImagePath(
-  path: string,
-  sourceDirectory: string,
-  hash: string
-): string {
-  return path
-    .replace(`${sourceDirectory}/${hash}`, BASE_IMAGES_DIRECTORY)
-    .replace(`${NEW_IMAGE_NAME}.png`, `${BASE_IMAGE_NAME}.png`);
-}
